+++ Changes.md
@@ -2,6 +2,7 @@
 
 ### 1.7.2
 - FIXED: first_or_create will now apply dirty tracking to newly created fields.
+- FIXED: Properties of :array type will always return a Parse::CollectionProxy if their internal value is nil. The object will not be marked dirty until something is added to the array.
 - NEW: `Parse::Model.autosave_on_create` has been removed in favor of `first_or_create!`.
 - FIXED: Encoding a Parse::Object into JSON will remove any values that are `nil`
 which were not explicitly changed to that value.